\name{bigQF-package}
\alias{bigQF-package}
\alias{bigQF}
\docType{package}
\title{
\packageTitle{bigQF}
}
\description{
\packageDescription{bigQF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bigQF}
\packageIndices{bigQF}
This package computes tail probabilities for large quadratic forms, with the motivation being the SKAT test used in DNA sequence association studies.  The main function is \code{pQF}, but also see \code{\link{SKAT.matrixfree}} for fast use with sparse genotypes.

}
\author{
\packageAuthor{bigQF}

Maintainer: \packageMaintainer{bigQF}
}
\references{
Lumley et al. (forthcoming) Sequence kernel association tests for large sets of markers: tail probabilities for large quadratic forms

Nathan Halko, Per-Gunnar Martinsson, Joel A. Tropp (2010) Finding structure with randomness: Probabilistic algorithms for constructing approximate matrix decompositions. \url{http://arxiv.org/abs/0909.4061}.

Lee, S., with contributions from Larisa Miropolsky, and Wu, M. (2015). SKAT: SNP-Set (Sequence) Kernel Association Test. R package version 1.1.2.

Lee, S., Wu, M. C., Cai, T., Li, Y., Boehnke, M., and Lin, X. (2011). Rare-variant association testing for sequencing data with the sequence kernel association test. American Journal of Human Genetics, 89:82-93.
}

